Separate validator.rs into traits.rs and types.rs
Many files use types defined by the coordinator only, so it's better to
separate the file into two. We may think of moving types.rs into codechain-types
crate as well.
