+++ core/src/block.rs
@@ -18,7 +18,8 @@ use crate::consensus::{ConsensusEngine, Evidence};
 use crate::error::{BlockError, Error};
 use ccrypto::BLAKE_NULL_RLP;
 use ckey::Address;
-use coordinator::validator::{BlockExecutor, Event, Header as PreHeader, Transaction, VerifiedCrime};
+use coordinator::traits::BlockExecutor;
+use coordinator::types::{Event, Header as PreHeader, Transaction, VerifiedCrime};
 use cstate::{StateDB, StateError, StateWithCache, TopLevelState};
 use ctypes::header::{Header, Seal};
 use ctypes::util::unexpected::Mismatch;
