+++ core/src/miner/mem_pool_types.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-use coordinator::validator::TransactionWithMetadata;
+use coordinator::types::TransactionWithMetadata;
 use ctypes::{BlockNumber, TxHash};
 use std::collections::HashMap;
